rename the directory, iij to iijpta.
