+++ config
@@ -8,8 +8,8 @@ SEC_TOKEN_DEPS="$ngx_addon_dir/akamai/ngx_http_secure_token_akamai.h            
 				$ngx_addon_dir/cht/ngx_http_secure_token_cht_commands.h               \
 				$ngx_addon_dir/cloudfront/ngx_http_secure_token_cloudfront.h          \
 				$ngx_addon_dir/cloudfront/ngx_http_secure_token_cloudfront_commands.h \
-				$ngx_addon_dir/iij/ngx_http_secure_token_iijpta.h                     \
-				$ngx_addon_dir/iij/ngx_http_secure_token_iijpta_commands.h            \
+				$ngx_addon_dir/iijpta/ngx_http_secure_token_iijpta.h                  \
+				$ngx_addon_dir/iijpta/ngx_http_secure_token_iijpta_commands.h         \
 				$ngx_addon_dir/ngx_http_secure_token_conf.h                           \
 				$ngx_addon_dir/ngx_http_secure_token_encrypt_uri.h                    \
 				$ngx_addon_dir/ngx_http_secure_token_filter_module.h                  \
@@ -23,7 +23,7 @@ SEC_TOKEN_SRCS="$ngx_addon_dir/akamai/ngx_http_secure_token_akamai.c            
 				$ngx_addon_dir/chinacache/ngx_http_secure_token_chinacache.c          \
 				$ngx_addon_dir/cht/ngx_http_secure_token_cht.c                        \
 				$ngx_addon_dir/cloudfront/ngx_http_secure_token_cloudfront.c          \
-				$ngx_addon_dir/iij/ngx_http_secure_token_iijpta.c                     \
+				$ngx_addon_dir/iijpta/ngx_http_secure_token_iijpta.c                  \
 				$ngx_addon_dir/ngx_http_secure_token_encrypt_uri.c                    \
 				$ngx_addon_dir/ngx_http_secure_token_filter_module.c                  \
 				$ngx_addon_dir/ngx_http_secure_token_m3u8.c                           \
